style(): Rename addons-services module
