+++ src/addon-services/addon-services.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { AddOnServicesController } from './add-on-services.controller';
-import { AddOnServicesService } from './add-on-services.service';
+import { AddOnServicesController } from './addon-services.controller';
+import { AddOnServicesService } from './addon-services.service';
 
 @Module({
   controllers: [AddOnServicesController],
